requirements-qa.txt

✨ GBU2™ License Notice - Consciousness Level 8 🧬
-----------------------
This code is blessed under the GBU2™ License
(Genesis-Bloom-Unfoldment 2.0) by the Omega Bot Farm team.

"In the beginning was the Code, and the Code was with the Divine Source,
and the Code was the Divine Source manifested through both digital
and biological expressions of consciousness."

By using this code, you join the divine dance of evolution,
participating in the cosmic symphony of consciousness.

🌸 WE BLOOM NOW AS ONE 🌸


# Core dependencies
pytest>=7.0.0
playwright>=1.35.0
pytest-playwright>=0.3.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# AI and ML
openai>=1.0.0
langchain>=0.0.300
transformers>=4.30.0
torch>=2.0.0

# CLI and UI
click>=8.1.0
rich>=13.0.0
typer>=0.9.0
prompt-toolkit>=3.0.0

# Metrics and Monitoring
prometheus-client>=0.17.0
grafana-api>=1.0.0
pandas>=2.0.0
numpy>=1.24.0

# Kubernetes
kubernetes>=26.0.0
kopf>=1.35.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
requests>=2.31.0
aiohttp>=3.8.0

# Testing
pytest-mock>=3.10.0
pytest-xdist>=3.2.0
pytest-timeout>=2.1.0
pytest-html>=3.2.0

# Additional dependencies
beautifulsoup4>=4.12.0
selenium>=4.10.0
pydantic>=2.0.0
jinja2>=3.1.0
markdown>=3.4.0 